requirements.txt
discord.py==1.3.3
maubot==0.1.0b1
